refactor : Move LottoDrawResult to upper package
