+++ src/test/java/lotto/domain/LottoGameTest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.domain.dto.LottoDrawResult;
 import lotto.domain.dto.LottoPurchaseResult;
 import lotto.util.LottoListGenerator;
 import org.junit.jupiter.api.BeforeEach;
